Make altOp the official op
